Continued big setting update

+++ app/Http/Controllers/Admin/Settings/DepartmentController.php
@@ -3,11 +3,11 @@
 namespace App\Http\Controllers\Admin\Settings;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\Settings\DepartmentsUpdateRequest;
+use App\Http\Requests\Admin\Settings\DepartmentRequest;
 use App\Models\Department;
 use App\Services\DiscordService;
+use App\Services\ImageService;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 class DepartmentController extends Controller
@@ -19,33 +19,23 @@ class DepartmentController extends Controller
         return view('admin.settings.department.index', compact('departments'));
     }
 
-    public function store(Request $request)
+    public function store(DepartmentRequest $request)
     {
 
-        $create = [
-            'name' => $request->input('name'),
-            'initials' => $request->input('initials'),
-            'type' => $request->input('type'),
-            'slug' => Str::slug($request->input('name')),
-        ];
+        $data = $request->validated();
+        $data['slug'] = Str::slug($data['name']);
 
-        if (! in_array($request->logo->extension(), ['jpg', 'jpeg', 'png', 'gif'])) {
-            return redirect()->route('admin.settings.department.create')->with('alerts', [['message' => 'Invalid Image Type.', 'level' => 'error']]);
-        }
+        $data['logo'] = ImageService::saveFromUrl(
+            url: $data['image_url'],
+            folder: 'images/departments/',
+            prefix: $data['slug']
+        );
 
-        $path = $request->logo->storePubliclyAs('images/departments/', Str::slug($create['slug']).'.'.$request->logo->extension(), 'public');
+        unset($data['image_url']);
 
-        $url = url('storage/images/departments/'.Str::slug($create['slug']).'.'.$request->logo->extension());
+        Department::create($data);
 
-        $create['logo'] = $url;
-
-        if (get_setting('discord.useRoles.useDepartmentRoles')) {
-            $create['discord_role_id'] = $request->input('discord_role_id');
-        }
-
-        $department = Department::create($create);
-
-        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department Created.', 'level' => 'success']]);
+        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department created.', 'level' => 'success']]);
     }
 
     public function create()
@@ -58,64 +48,46 @@ class DepartmentController extends Controller
         return view('admin.settings.department.create', compact('discordRoles'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Department $department)
     {
-        $discord_roles = [];
+        $discordRoles = [];
         if (get_setting('discord.useRoles.useDepartmentRoles')) {
-            $discord_roles = (new DiscordService)->get_server_roles();
+            $discordRoles = DiscordService::getServerRoles();
         }
 
-        return view('admin.settings.department.edit', compact('discord_roles', 'department'));
+        return view('admin.settings.department.edit', compact('discordRoles', 'department'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(DepartmentsUpdateRequest $request, Department $department)
+    public function destroy(Request $request, Department $department)
     {
-        $update = [
-            'name' => $request->validated('name'),
-            'initials' => $request->validated('initials'),
-            'type' => $request->validated('type'),
-            'slug' => Str::slug($request->validated('name')),
-        ];
+        $confirm = $request->input('confirm');
 
-        if ($request->file('logo')) {
-            if (! in_array($request->logo->extension(), ['jpg', 'jpeg', 'png', 'gif'])) {
-                return redirect()->route('admin.settings.department.edit', $department->slug)->with('alerts', [['message' => 'Invalid Image Type.', 'level' => 'error']]);
-            }
-
-            $path = $request->logo->storePubliclyAs('images/departments/', Str::slug($request->validated('name')).'.'.$request->logo->extension(), 'public');
-
-            $url = url('storage/images/departments/'.Str::slug($request->validated('name')).'.'.$request->logo->extension());
-
-            $update['logo'] = $url;
+        if ($confirm !== $department->name) {
+            return redirect()->route('admin.settings.departments.edit', $department->slug)->with('alerts', [['message' => 'Department delete confirm check didn\'t match.', 'level' => 'error']]);
         }
 
-        if (get_setting('discord.useRoles.useDepartmentRoles')) {
-            $department->update(['discord_role_id' => $request->input('discord_role_id')]);
-            $update['discord_role_id'] = $request->input('discord_role_id');
-        }
-
-        $department->update($update);
-
-        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department Saved.', 'level' => 'success']]);
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Department $department)
-    {
-        $file_name = explode('/', $department->logo);
-        $file = end($file_name);
-        Storage::disk('public')->delete('images/departments/'.$file);
-
+        ImageService::deleteFromUrl($department->logo);
+        $department->update(['logo' => null, 'discord_role_id' => null]);
         $department->delete();
 
-        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department Deleted.', 'level' => 'success']]);
+        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department deleted.', 'level' => 'success']]);
+    }
+
+    public function update(DepartmentRequest $request, Department $department)
+    {
+        $data = $request->validated();
+        $data['slug'] = Str::slug($data['name']);
+
+        $data['logo'] = ImageService::saveFromUrl(
+            url: $data['image_url'],
+            folder: 'images/departments/',
+            prefix: $data['slug']
+        );
+
+        unset($data['image_url']);
+
+        $department->update($data);
+
+        return redirect()->route('admin.settings.departments.index')->with('alerts', [['message' => 'Department saved.', 'level' => 'success']]);
     }
 }
